TelegramBot/requirements.txt
python-telegram-bot>=21.8
requests>=2.31.0
chatgpt-md-converter>=0.1.1
ddgs>=9.0.0
